youtubeaudiobackend/requirement.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
pafy==0.5.5
pytube==15.0.0
pytubefix==6.13.1
sqlparse==0.5.1
